Handle CONC properly by not including newlines.
From the GEDCOM 5.5 Spec
(http://homepages.rootsweb.ancestry.com/~pmcbride/gedcom/55gcappa.htm)

CONC {CONCATENATION}: = An indicator that additional data belongs to the
superior value. The information from the CONC value is to be connected to
the value of the superior preceding line without a space and without a
carriage return and/or new line character. Values that are split for a
CONC tag must always be split at a non-space. If the value is split on a
space the space will be lost when concatenation takes place. This is
because of the treatment that spaces get as a GEDCOM delimiter, many
GEDCOM values are trimmed of trailing spaces and some systems look for
the first non-space starting after the tag to determine the beginning of
the value.

CONT {CONTINUED}: = An indicator that additional data belongs to the
superior value. The information from the CONT value is to be connected to
the value of the superior preceding line with a carriage return and/or
new line character. Leading spaces could be important to the formatting
of the resultant text. When importing values from CONT lines the reader
should assume only one delimiter character following the CONT tag. Assume
that the rest of the leading spaces are to be a part of the value.

Signed-off-by: Tim Shadel <github@timshadel.com>